+++ Controllers/Production/ProductionOrderController.cs
@@ -17,79 +17,68 @@ public class ProductionOrderController : ControllerBase
         _service = service;
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateProductionOrderDto dto)
     {
 
-/*         if (!User.HasClaim("permission", "CREATE:PRODUCTION_ORDER"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "CREATE:PRODUCTION_ORDERS")) return Forbid();
 
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll([FromQuery] string? receiptNumber, [FromQuery] int? product, [FromQuery] DateTime? startDate, [FromQuery] DateTime? endDate)
     {
 
-/*         if (!User.HasClaim("permission", "VIEW:PRODUCTION_ORDER"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "VIEW:PRODUCTION_ORDERS")) return Forbid();
 
         var result = await _service.GetAllAsync(receiptNumber, product, startDate, endDate);
 
-        if (result == null)
-            return NotFound();
+        if (result == null) return NotFound();
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
 
-/*         if (!User.HasClaim("permission", "VIEW:PRODUCTION_ORDER"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "VIEW:PRODUCTION_ORDERS")) return Forbid();
 
         var result = await _service.GetByIdAsync(id);
 
-        if (result == null)
-            return NotFound();
+        if (result == null) return NotFound();
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] ProductionOrderDto dto)
     {
-
-/*         if (!User.HasClaim("permission", "UPDATE:PRODUCTION_ORDER"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "EDIT:PRODUCTION_ORDERS")) return Forbid();
 
         var result = await _service.UpdateAsync(id, dto);
 
-        if (!result)
-            NotFound();
+        if (!result) return NotFound();
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
-
-/*         if (!User.HasClaim("permission", "DELETE:PRODUCTION_ORDER"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "DELETE:PRODUCTION_ORDERS")) return Forbid();
 
         var result = await _service.DeleteAsync(id);
 
-        if (!result)
-            NotFound();
+        if (!result) return NotFound();
 
         return Ok(result);
     }
