+++ Controllers/Sales/CustomerController.cs
@@ -17,26 +17,26 @@ public class CustomerController : ControllerBase
         _service = service;
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPost]
     public async Task<IActionResult> Create([FromBody] CreateCustomerDto dto)
     {
 
-        /* if (!User.HasClaim("permission", "CREATE:SUPPLIERS"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "CREATE:CUSTOMERS"))
+            return Forbid();
 
         var result = await _service.CreateAsync(dto);
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpGet]
     public async Task<IActionResult> GetAll()
     {
 
-        /* if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "VIEW:CUSTOMERS"))
+            return Forbid();
 
         var result = await _service.GetAllAsync();
 
@@ -46,50 +46,44 @@ public class CustomerController : ControllerBase
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpGet("{id}")]
     public async Task<IActionResult> GetById(int id)
     {
 
-        /* if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "VIEW:CUSTOMERS")) return Forbid();
 
         var result = await _service.GetByIdAsync(id);
 
-        if (result == null)
-            return NotFound();
+        if (result == null) return NotFound();
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpPut("{id}")]
     public async Task<IActionResult> Update(int id, [FromBody] UpdateCustomerDto dto)
     {
 
-        /* if (!User.HasClaim("permission", "VIEW:SUPPLIERS"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "VIEW:CUSTOMERS")) return Forbid();
 
         var result = await _service.UpdateAsync(id, dto);
 
-        if (!result)
-            NotFound();
+        if (!result) return NotFound();
 
         return Ok(result);
     }
 
-    //[Authorize]
+    [Authorize]
     [HttpDelete("{id}")]
     public async Task<IActionResult> Delete(int id)
     {
 
-        /* if (!User.HasClaim("permission", "DELETE:SUPPLIERS"))
-            return Forbid(); */
+        if (!User.HasClaim("permission", "DELETE:CUSTOMERS")) return Forbid();
 
         var result = await _service.DeleteAsync(id);
 
-        if (!result)
-            NotFound();
+        if (!result) return NotFound();
 
         return Ok(result);
     }
